+++ src/ArtifactoryUploader/PackageUploadHelper.cs
@@ -997,8 +997,7 @@ namespace LCT.ArtifactoryUploader
             }
 
             if (SetWarningCode)
-            {
-                CommonHelper.PublishFilesToArtifact();
+            {                
                 Environment.ExitCode = 2;
                 Logger.Debug("Setting ExitCode to 2");
             }
