+++ src/LCT.PackageIdentifier/Program.cs
@@ -117,7 +117,7 @@ namespace LCT.PackageIdentifier
             Logger.Logger.Log(null, Level.Notice, $"End of Package Identifier execution : {DateTime.Now}\n", null);
 
             // publish logs and bom file to pipeline artifact
-            CommonHelper.PublishFilesToArtifact();
+            PublishPipelineArtifacts.PublishFilesToArtifact();
 
         }
